requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
libmagic==1.0
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2024.1
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
